requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
diffusers
transformers
accelerate
gTTS
moviepy
numpy
imageio-ffmpeg
Pillow==10.1.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
skia-python
uharfbuzz
